requirements.txt
flask==3.0.0
flask-jwt-extended==4.5.3
httpx==0.24.1
jsonify==0.5
python-dotenv==1.0.0
psycopg2==2.9.9
urllib3==1.26.6
workos==1.2.0
